Implement review item question flow from the user's perspective
