+++ frontend/models/request/review_item.go
@@ -1,0 +1,7 @@
+package request
+
+type SubmitReviewItemQuestionForm struct {
+	SingleChoiceValue   string   `form:"single-choice-value"`
+	MultipleChoiceValue []string `form:"multiple-choice-value"`
+	TrueOrFalseValue    bool     `form:"true-or-false-value"`
+}
